requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SQLAlchemy>=3.0
python-dateutil>=2.8
SQLAlchemy>=1.4
